backend/requirements.txt
fastapi
langchain
langchain-core
langchain-perplexity
langchain-google-genai
uvicorn
pydantic
python-box
ensure
python-dotenv
motor
passlib
pytest
pytest-asyncio
pydantic-settings
pydantic[email]
python-jose
bcrypt
python-jose[cryptography]
email-validator
pyjwt
sendgrid
streamlit
pillow
#google
#google-generativeai
#IPython